requirements.txt
Pillow>=10.4.0
PyYAML>=6.0.2
piexif>=1.1.3

# first start program - initial 
# python -m venv .venv
# .venv\Scripts\pip install -r requirements.txt


# run program
# .venv\Scripts\python main.py
